requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
slugify==0.0.1
sqlparse==0.5.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.18.3
